HTBHF-2227 Moved manual address postcode validation to common address folder.
