+++ src/web/routes/application/address/manual-address/validate.js
@@ -1,12 +1,7 @@
 const { check } = require('express-validator')
 const { translateValidationMessage } = require('../../common/translate-validation-message')
+const { UK_POSTCODE_PATTERN } = require('../constants')
 
-/*
- * Regex for matching UK postcodes matching BS7666 format.
- * see https://www.gov.uk/government/publications/bulk-data-transfer-for-sponsors-xml-schema  The format is in the file BulkDataCommon-v2.1.xsd
- * see https://stackoverflow.com/questions/164979/uk-postcode-regex-comprehensive
- */
-const UK_POSTCODE_PATTERN = /^([Gg][Ii][Rr] 0[Aa]{2})$|^((([A-Za-z][0-9]{1,2})|(([A-Za-z][A-Ha-hJ-Yj-y][0-9]{1,2})|(([A-Za-z][0-9][A-Za-z])|([A-Za-z][A-Ha-hJ-Yj-y][0-9][A-Za-z]?))))\s?[0-9][A-Za-z]{2})$/
 const ADDRESS_LINE_MAX_LENGTH = 500
 
 const validate = () => [
@@ -40,6 +35,5 @@ const validate = () => [
 ]
 
 module.exports = {
-  validate,
-  UK_POSTCODE_PATTERN
+  validate
 }
